+++ src/agent/agent.py
@@ -4,9 +4,9 @@ from dotenv import load_dotenv
 from queue import Queue
 from src.agent.providers.model_provider import ModelProvider
 from src.agent.providers.search_provider import SearchProvider
-from src.agent.sentient_chat.implementation.default_response_handler import DefaultResponseHandler
-from src.agent.sentient_chat.implementation.default_hook import DefaultHook
-from src.agent.sentient_chat.interface.identity import Identity
+from src.agent.sentient_agent.implementation.default_response_handler import DefaultResponseHandler
+from src.agent.sentient_agent.implementation.default_hook import DefaultHook
+from src.agent.sentient_agent.interface.identity import Identity
 from typing import Iterator
