+++ src/agent/sentient_agent/interface/stream_event_emitter.py
